Add apiCalls file to clean up App
